Fix all audit issues: prefix/postfix f64 return types, test_results.json accuracy, Cargo.toml repo URL, test hash paths, .gitignore, Dockerfile, README speedup claims, fuzz log documentation, DECISIONS.md Bug Catcher clarification, adapter test coverage (gotoh/nw/sw), p99 latency, tests/port/ directory

+++ src/algorithms/sequence/lcsseq.rs
@@ -13,13 +13,13 @@ impl LCSSeq {
         Self
     }
 
-    pub fn similarity(&self, s1: &str, s2: &str) -> usize {
-        self.lcs(s1, s2).chars().count()
+    pub fn similarity(&self, s1: &str, s2: &str) -> f64 {
+        self.lcs(s1, s2).chars().count() as f64
     }
 
     pub fn distance(&self, s1: &str, s2: &str) -> f64 {
-        let max_len = max(s1.chars().count(), s2.chars().count());
-        (max_len - self.similarity(s1, s2)) as f64
+        let max_len = max(s1.chars().count(), s2.chars().count()) as f64;
+        max_len - self.similarity(s1, s2)
     }
 
     fn lcs(&self, s1: &str, s2: &str) -> String {
